sbtool-sign-kernel: handle kernels without signing keys configured
We assume that the kernel is built with a signing key but that
assumption will likely be invalid on many systems.  If a signing key
is unspecified and we can find a config but module signing is disabled.
If a signing key is unspecified and we can't find a config, exit with
an error.