scripts: move messages around to better handle skipped enrollment
Skipping enrollment and skipping signing/install need to be handled
differently.  We can have each scripts exit with retval=2 if they
have skipped the operation and the end state means the work has not
been performed.  In order to properly report what's happened, the
message printing needs to be moved out of installkernel and into
the called scripts.